+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * [Change] Add localizations to the missed strings in the mod.
 * [Enhancement] Add standby dialog for the action that can take much time to complete.
 * [Enhancement] Add RU localization.
+* [Enhancement] Support "global strings". They can be used multiple times across the mod.
 * [Enhancement #12] Properly handle color values.
 * [Enhancement #13] Exclude the stock localization tags from the exported files.
 * [Enhancement #14] Add ability to switch the current game language in runtime.